Random picking of a record works!
Fixed a bug in both the pseudo-code and in the code,
which prevented from picking the right index of a record corresponding
to a randomly generated number. The right index is the one in
correspondence of which the random number is not greater than the sum of
weights (at that index), but it is greater than the previous sum (at the
index before).

Assert conditions added, to check some functions's input arguments.