Printing of probabilities and frequencies to stdout
Records'probabilities and frequencies are shown on the stdout. One can
see that, if the number of iterations gets higher and higher, then
frequencies get closer and closer to the corresponding probabilities.
This proves that the algorithm works fine.
Fixed a bug on update of indexes of array of cum. weights, which used to
cause a segmentation fault.